Move modulestore django settings migration code down into the modulestore layer so it works for all django-based callers.
